Update error handling

+++ delete.php
@@ -12,20 +12,28 @@ if (isset($_POST['submit'])) {
 		die('Failed to connect to MySQL: '.mysqli_connect_error());
 	}
 
-    $ProductName = $_POST['ProductName'];
+    //Test if table exists
+    $res = mysqli_query($conn, "SHOW TABLES LIKE 'Products'");
+    if (mysqli_num_rows($res) <= 0) {
+        echo "<h2>Catalog is empty</h2>";
+    } else { 
 
-    if ($stmt = mysqli_prepare($conn, "DELETE FROM Products WHERE ProductName = ?")) {
-        mysqli_stmt_bind_param($stmt, 's', $ProductName);
-        mysqli_stmt_execute($stmt);
-        if (mysqli_stmt_affected_rows($stmt) == 0) {
-            echo "<h2>Product \"$ProductName\" was not found in the catalog.</h2>";
-        }
-        else {
-            echo "<h2>Product \"$ProductName\" has been removed from the catalog.</h2>";
-        }
+        //Delete data
+        $ProductName = $_POST['ProductName'];
 
-        mysqli_stmt_close($stmt);
-        
+        if ($stmt = mysqli_prepare($conn, "DELETE FROM Products WHERE ProductName = ?")) {
+            mysqli_stmt_bind_param($stmt, 's', $ProductName);
+            mysqli_stmt_execute($stmt);
+            if (mysqli_stmt_affected_rows($stmt) == 0) {
+                echo "<h2>Product \"$ProductName\" was not found in the catalog.</h2>";
+            }
+            else {
+                echo "<h2>Product \"$ProductName\" has been removed from the catalog.</h2>";
+            }
+
+            mysqli_stmt_close($stmt);
+            
+        }
     } 
 
     //Close the connection
